requirements.txt
pycryptodome
hashlib
cryptography